requirements.txt
python-dotenv
langchain
langchain-docling
transformers
huggingface-hub
torch
docling
transformers
torch
Pillow

